Fixing the bug that FLIR is not activatable after a few hours of play time (#424)
Inside the different FLIR scripts there was a mixture of 'time' variable and 'cba_missionTime'. These two variables are not comparable! After a few hours of play time they can deviate in minutes or even hours (!) Now FLIR consequently uses 'cba_missionTime' for announcing the boot time as well as for the check when the FLIR system has been booted.

Co-authored-by: Schmitt <tommy.ludwig@gmx.de>